Bugfixes while testing

+++ Vitae/Vitae/Areas/Manage/Pages/Settings/Index.cshtml.cs
@@ -86,91 +86,94 @@ namespace Vitae.Areas.Manage.Pages.Settings
 
         public async Task<IActionResult> OnPostAsync()
         {
-            var existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID);
-            var firstLanguage = existingCurriculumLanguages.Single(e => e.Order == 0).Language;
-            var deletedLanguageCodes = existingCurriculumLanguages.ToList().Select(e => e.Language.LanguageCode).Where(lc => !Setting.SettingItems.Any(s => s.FormerLanguageCode == lc));
-            var ignoreOrder = new List<int>();
-
-            // Delete
-            foreach (var deletedLanguageCode in deletedLanguageCodes)
+            if (ModelState.IsValid)
             {
-                vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(e => e.Language.LanguageCode == deletedLanguageCode));
+                var existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID);
+                var firstLanguage = existingCurriculumLanguages.Single(e => e.Order == 0).Language;
+                var deletedLanguageCodes = existingCurriculumLanguages.ToList().Select(e => e.Language.LanguageCode).Where(lc => !Setting.SettingItems.Any(s => s.FormerLanguageCode == lc));
+                var ignoreOrder = new List<int>();
+
+                // Delete
+                foreach (var deletedLanguageCode in deletedLanguageCodes)
+                {
+                    vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(e => e.Language.LanguageCode == deletedLanguageCode));
+                    await vitaeContext.SaveChangesAsync();
+
+                    await repository.DeleteItemsFromCurriculumLanguageAsync(curriculumID, deletedLanguageCode);
+                }
+
+                // Order
+                for (int i = 0; i < vitaeContext.CurriculumLanguages.Count(cl => cl.CurriculumID == curriculumID); i++)
+                {
+                    vitaeContext.CurriculumLanguages.Where(cl => cl.CurriculumID == curriculumID).OrderBy(c => c.Order).ToList()[i].Order = i;
+                }
                 await vitaeContext.SaveChangesAsync();
 
-                await repository.DeleteItemsFromCurriculumLanguageAsync(curriculumID, deletedLanguageCode);
-            }
+                existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID);
 
-            // Order
-            for (int i = 0; i < vitaeContext.CurriculumLanguages.Count(cl => cl.CurriculumID == curriculumID); i++)
-            {
-                vitaeContext.CurriculumLanguages.Where(cl => cl.CurriculumID == curriculumID).OrderBy(c => c.Order).ToList()[i].Order = i;
-            }
-            await vitaeContext.SaveChangesAsync();
-
-            existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID);
-
-            // Change / Add
-            for (int i = 0; i < Setting.CurriculumLanguages.Count; i++)
-            {
-                var existingLanguage = existingCurriculumLanguages.SingleOrDefault(e => e.Order == i)?.Language;
-                var formerLanguageCode = Setting.SettingItems[i].FormerLanguageCode;
-                var newLanguage = Setting.CurriculumLanguages[i];
-
-                if(formerLanguageCode != null && formerLanguageCode != newLanguage.LanguageCode)
+                // Change / Add
+                for (int i = 0; i < Setting.CurriculumLanguages.Count; i++)
                 {
-                    var curriculum = existingCurriculumLanguages.Single(e => e.Order == i).Curriculum;
+                    var existingLanguage = existingCurriculumLanguages.SingleOrDefault(e => e.Order == i)?.Language;
+                    var formerLanguageCode = Setting.SettingItems[i].FormerLanguageCode;
+                    var newLanguage = Setting.CurriculumLanguages[i];
 
-                    if (!vitaeContext.CurriculumLanguages.Any(c => c.Language.LanguageCode == newLanguage.LanguageCode))
+                    if (formerLanguageCode != null && formerLanguageCode != newLanguage.LanguageCode)
                     {
-                        // The language code has changed
-                        vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(o => o.Order == i));
-                        AddCurriculumLanguage(curriculum, vitaeContext.Languages.Single(l => l.LanguageCode == newLanguage.LanguageCode), i);
+                        var curriculum = existingCurriculumLanguages.Single(e => e.Order == i).Curriculum;
 
-                        // Change items
-                        if (Setting.SettingItems[i].NrOfItems > 0)
+                        if (!vitaeContext.CurriculumLanguages.Any(c => c.Language.LanguageCode == newLanguage.LanguageCode))
                         {
-                            await repository.MoveItemsFromCurriculumLanguageAsync(curriculumID, existingLanguage.LanguageCode, newLanguage.LanguageCode, copy: false);
+                            // The language code has changed
+                            vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(o => o.Order == i));
+                            AddCurriculumLanguage(curriculum, vitaeContext.Languages.Single(l => l.LanguageCode == newLanguage.LanguageCode), i);
+
+                            // Change items
+                            if (Setting.SettingItems[i].NrOfItems > 0)
+                            {
+                                await repository.MoveItemsFromCurriculumLanguageAsync(curriculumID, existingLanguage.LanguageCode, newLanguage.LanguageCode, copy: false);
+                            }
                         }
+                        else if (!ignoreOrder.Contains(i))
+                        {
+                            // Language switch occured
+                            var newOrder = Setting.CurriculumLanguages.Single(c => c.LanguageCode == newLanguage.LanguageCode).Order;
+                            var oldOrder = existingCurriculumLanguages.Single(c => c.Language.LanguageCode == newLanguage.LanguageCode).Order;
+                            existingCurriculumLanguages.Single(c => c.Order == newOrder).Order = oldOrder;
+                            existingCurriculumLanguages.Single(c => c.Language.LanguageCode == newLanguage.LanguageCode).Order = newOrder;
+
+                            ignoreOrder.Add(oldOrder);
+
+                        }
+                        curriculum.LastUpdated = DateTime.Now;
+                        await vitaeContext.SaveChangesAsync();
+
+                        // Update View Model
+                        Setting.SettingItems[i].FormerLanguageCode = newLanguage.LanguageCode;
+                        ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.FormerLanguageCode)}", new ValueProviderResult(newLanguage.LanguageCode, CultureInfo.InvariantCulture));
+
                     }
-                    else if(!ignoreOrder.Contains(i))
+                    else if (existingLanguage == null)
                     {
-                        // Language switch occured
-                        var newOrder = Setting.CurriculumLanguages.Single(c => c.LanguageCode == newLanguage.LanguageCode).Order;
-                        var oldOrder = existingCurriculumLanguages.Single(c => c.Language.LanguageCode == newLanguage.LanguageCode).Order;
-                        existingCurriculumLanguages.Single(c => c.Order == newOrder).Order = oldOrder;
-                        existingCurriculumLanguages.Single(c => c.Language.LanguageCode == newLanguage.LanguageCode).Order = newOrder;
+                        // There's a new langauge at the end
+                        AddCurriculumLanguage(existingCurriculumLanguages.Single(e => e.Order == 0).Curriculum, vitaeContext.Languages.Single(l => l.LanguageCode == newLanguage.LanguageCode), i);
+                        await vitaeContext.SaveChangesAsync();
 
-                        ignoreOrder.Add(oldOrder);
+                        if (Setting.SettingItems[i].Copy)
+                        {
+                            // Copy from first language
+                            await repository.MoveItemsFromCurriculumLanguageAsync(curriculumID, firstLanguage.LanguageCode, newLanguage.LanguageCode, copy: true);
+                        }
 
+                        // Update View Model
+                        var nrOfItems = (await repository.CountItemsFromCurriculumAsync(curriculumID, newLanguage.LanguageCode)).Single().Value.Sum(i => i.Value);
+                        Setting.SettingItems[i].Copy = false;
+                        Setting.SettingItems[i].FormerLanguageCode = newLanguage.LanguageCode;
+                        Setting.SettingItems[i].NrOfItems = nrOfItems;
+                        ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.Copy)}", new ValueProviderResult("false", CultureInfo.InvariantCulture));
+                        ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.FormerLanguageCode)}", new ValueProviderResult(newLanguage.LanguageCode, CultureInfo.InvariantCulture));
+                        ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.NrOfItems)}", new ValueProviderResult(nrOfItems.ToString(), CultureInfo.InvariantCulture));
                     }
-                    curriculum.LastUpdated = DateTime.Now;
-                    await vitaeContext.SaveChangesAsync();
-
-                    // Update View Model
-                    Setting.SettingItems[i].FormerLanguageCode = newLanguage.LanguageCode;
-                    ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.FormerLanguageCode)}", new ValueProviderResult(newLanguage.LanguageCode, CultureInfo.InvariantCulture));
-
-                }
-                else if (existingLanguage == null)
-                {
-                    // There's a new langauge at the end
-                    AddCurriculumLanguage(existingCurriculumLanguages.Single(e => e.Order == 0).Curriculum, vitaeContext.Languages.Single(l => l.LanguageCode == newLanguage.LanguageCode), i);
-                    await vitaeContext.SaveChangesAsync();
-
-                    if (Setting.SettingItems[i].Copy)
-                    {
-                        // Copy from first language
-                        await repository.MoveItemsFromCurriculumLanguageAsync(curriculumID, firstLanguage.LanguageCode, newLanguage.LanguageCode, copy: true);
-                    }
-
-                    // Update View Model
-                    var nrOfItems = (await repository.CountItemsFromCurriculumAsync(curriculumID, newLanguage.LanguageCode)).Single().Value.Sum(i => i.Value);
-                    Setting.SettingItems[i].Copy = false;
-                    Setting.SettingItems[i].FormerLanguageCode = newLanguage.LanguageCode;
-                    Setting.SettingItems[i].NrOfItems = nrOfItems;
-                    ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.Copy)}", new ValueProviderResult("false", CultureInfo.InvariantCulture));
-                    ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.FormerLanguageCode)}", new ValueProviderResult(newLanguage.LanguageCode, CultureInfo.InvariantCulture));
-                    ModelState.SetModelValue($"{nameof(Setting)}.{nameof(Setting.SettingItems)}[{i}].{nameof(SettingItemVM.NrOfItems)}", new ValueProviderResult(nrOfItems.ToString(), CultureInfo.InvariantCulture));
                 }
             }
 
@@ -181,43 +184,50 @@ namespace Vitae.Areas.Manage.Pages.Settings
 
         public async Task<RedirectResult> OnPostDeleteAccountAsync()
         {
-            // Delete
-            var existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID).ToList();
-            await DeleteLanguagesAsync(existingCurriculumLanguages, existingCurriculumLanguages.Select(e => e.Language.LanguageCode));
-
-            vitaeContext.CurriculumLanguages.RemoveRange(vitaeContext.CurriculumLanguages.Include(c => c.Language).Where(c => c.CurriculumID == curriculumID));
-            vitaeContext.Publications.RemoveRange(vitaeContext.Publications.Include(p => p.Curriculum).Where(p => p.Curriculum.CurriculumID == curriculumID));
-            var curriculum = await repository.GetCurriculumAsync<PersonalDetail>(curriculumID);
-
-            // Remove core values
-            if (curriculum.PersonalDetails.Count == 1)
+            if (ModelState.IsValid)
             {
-                curriculum.PersonalDetails.Single().Children.ToList().ForEach(c => vitaeContext.Entry(c).State = EntityState.Deleted);
-                curriculum.PersonalDetails.Single().PersonCountries.ToList().ForEach(p => vitaeContext.Entry(p).State = EntityState.Deleted);
-                vitaeContext.Entry(curriculum.PersonalDetails.Single()).State = EntityState.Deleted;
+                // Delete
+                var existingCurriculumLanguages = vitaeContext.CurriculumLanguages.Include(c => c.Language).Include(c => c.Curriculum).Where(cl => cl.CurriculumID == curriculumID).ToList();
+                await DeleteLanguagesAsync(existingCurriculumLanguages, existingCurriculumLanguages.Select(e => e.Language.LanguageCode));
+
+                vitaeContext.CurriculumLanguages.RemoveRange(vitaeContext.CurriculumLanguages.Include(c => c.Language).Where(c => c.CurriculumID == curriculumID));
+                vitaeContext.Publications.RemoveRange(vitaeContext.Publications.Include(p => p.Curriculum).Where(p => p.Curriculum.CurriculumID == curriculumID));
+                var curriculum = await repository.GetCurriculumAsync<PersonalDetail>(curriculumID);
+
+                // Remove core values
+                if (curriculum.PersonalDetails.Count == 1)
+                {
+                    curriculum.PersonalDetails.Single().Children.ToList().ForEach(c => vitaeContext.Entry(c).State = EntityState.Deleted);
+                    curriculum.PersonalDetails.Single().PersonCountries.ToList().ForEach(p => vitaeContext.Entry(p).State = EntityState.Deleted);
+                    vitaeContext.Entry(curriculum.PersonalDetails.Single()).State = EntityState.Deleted;
+                }
+                vitaeContext.Entry(vitaeContext.Curriculums.Single(c => c.CurriculumID == curriculumID)).State = EntityState.Deleted;
+
+                await repository.LogAsync(curriculumID, null, LogArea.Delete, LogLevel.Information, CodeHelper.GetCalledUri(httpContext), CodeHelper.GetUserAgent(httpContext), requestCulture.RequestCulture.UICulture.Name, httpContext.Connection.RemoteIpAddress.ToString());
+
+                await vitaeContext.SaveChangesAsync();
+
+                // Identity context
+                var user = await userManager.GetUserAsync(User);
+                identityContext.Users.Remove(identityContext.Users.Single(u => u.Id == user.Id));
+
+                await identityContext.SaveChangesAsync();
+
+                // Send mail
+                var email = await userManager.GetEmailAsync(user);
+                var bodyText = await CodeHelper.GetMailBodyTextAsync(SharedResource.AccountDeleted, email, new Tuple<string, string, string>(SharedResource.MailBye3, Globals.APPLICATION_URL, SharedResource.ClickingHere), false, SharedResource.Goodbye);
+                var logoStream = CodeHelper.GetLogoStream(Globals.LOGO);
+                var message = new Message(new string[] { email }, SharedResource.AccountDeleted, bodyText, new FormFileCollection() { new FormFile(logoStream, 0, logoStream.Length, "image/png", "logo") });
+                await emailSender.SendEmailAsync(message);
+
+                await signInManager.SignOutAsync();
+
+                return Redirect("/");
             }
-            vitaeContext.Entry(vitaeContext.Curriculums.Single(c => c.CurriculumID == curriculumID)).State = EntityState.Deleted;
-
-            await repository.LogAsync(curriculumID, null, LogArea.Delete, LogLevel.Information, CodeHelper.GetCalledUri(httpContext), CodeHelper.GetUserAgent(httpContext), requestCulture.RequestCulture.UICulture.Name, httpContext.Connection.RemoteIpAddress.ToString());
-
-            await vitaeContext.SaveChangesAsync();
-
-            // Identity context
-            var user = await userManager.GetUserAsync(User);
-            identityContext.Users.Remove(identityContext.Users.Single(u => u.Id == user.Id));
-
-            await identityContext.SaveChangesAsync();
-
-            // Send mail
-            var email = await userManager.GetEmailAsync(user);
-            var bodyText = await CodeHelper.GetMailBodyTextAsync(SharedResource.AccountDeleted, email, new Tuple<string, string, string>(SharedResource.MailBye3, Globals.APPLICATION_URL, SharedResource.ClickingHere), false, SharedResource.Goodbye);
-            var logoStream = CodeHelper.GetLogoStream(Globals.LOGO);
-            var message = new Message(new string[] { email }, SharedResource.AccountDeleted, bodyText, new FormFileCollection() { new FormFile(logoStream, 0, logoStream.Length, "image/png", "logo") });
-            await emailSender.SendEmailAsync(message);
-
-            await signInManager.SignOutAsync();
-
-            return Redirect("/");
+            else
+            {
+                return Redirect($"/{nameof(Manage)}/{nameof(Settings)}");
+            }
         }
 
         #endregion
@@ -262,15 +272,18 @@ namespace Vitae.Areas.Manage.Pages.Settings
 
         private async Task DeleteLanguagesAsync(IEnumerable<CurriculumLanguage> existingCurriculumLanguages, IEnumerable<string> deletedLanguageCodes)
         {
-            // Delete
-            foreach (var deletedLanguageCode in deletedLanguageCodes)
+            if (ModelState.IsValid)
             {
-                vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(e => e.Language.LanguageCode == deletedLanguageCode));
-                vitaeContext.Publications.RemoveRange(vitaeContext.Publications.Include(p => p.CurriculumLanguage).Where(p => p.Curriculum.CurriculumID == curriculumID && p.CurriculumLanguage.LanguageCode == deletedLanguageCode));
-                
-                await vitaeContext.SaveChangesAsync();
+                // Delete
+                foreach (var deletedLanguageCode in deletedLanguageCodes)
+                {
+                    vitaeContext.CurriculumLanguages.Remove(existingCurriculumLanguages.Single(e => e.Language.LanguageCode == deletedLanguageCode));
+                    vitaeContext.Publications.RemoveRange(vitaeContext.Publications.Include(p => p.CurriculumLanguage).Where(p => p.Curriculum.CurriculumID == curriculumID && p.CurriculumLanguage.LanguageCode == deletedLanguageCode));
 
-                await repository.DeleteItemsFromCurriculumLanguageAsync(curriculumID, deletedLanguageCode);
+                    await vitaeContext.SaveChangesAsync();
+
+                    await repository.DeleteItemsFromCurriculumLanguageAsync(curriculumID, deletedLanguageCode);
+                }
             }
         }
